Fix indentation of "fi" inserted for an indented "if" by the shell script Ctrl-k mapping.
Previously the "fi" inserted was not indented enough.

The changes to plugin/code_block_markers.vim undo the c4eee25bf0a661 and 4026620bbce8e6 commits. Those commits appear to have been really ugly patches that didn't fix the root causes of the problem/s; We should be able to rely on automatic indentation when adding "fi" (or "}"). Now, with Vim 8.1-2269, all tests pass with the changes made in those commits removed (and the updated "if (sh)" test would fail if they were still present).